pyproject.toml
[tool.poetry]
name = "quiltcore"
version = "0.4.1"
description = "low-level plubming to read/write Quilt packages"
authors = ["Ernest Prabhakar <ernest@quiltdata.io>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
jsonlines = "1.2.0"
multiformats = "^0.2.1"
pandas = "^2.0.2"
pyarrow = "^13.0.0"
pandas-stubs = "^2.0.2.230605"
quilt3 = "^5.3.1"
s3fs = "^2023.6.0"
typing-extensions = "^4.6.3"
tzlocal = "^5.0.1"
un-yaml = ">=0.3.1"
# un-yaml = {git = "https://github.com/data-yaml/un-yaml.git", rev = "main"}
universal-pathlib = "^0.1.3"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.0.0"
mypy = "^1.4.1"
pytest-watcher = "^0.3.4"
pytest-codeblocks = "^0.17.0"
pytest-cov = "^4.0.0"
types-pyyaml = "^6.0.12.10"
types-tzlocal = "^5.0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest-watcher]
now = true
delay = 0.5
runner = "pytest"
runner_args = []
patterns = ["*.py", "*.yaml", "Makefile", "pyproject.toml", "poetry.lock", "3210f808ac0467726439191eea3bd0a66ab261122ee55758367620fedc77fe08"]
ignore_patterns = []